+++ tests/ops/integration_tests/saas/request_override/test_mailchimp_override_task.py
@@ -29,7 +29,7 @@ as the standard Mailchimp config.
 
 @pytest.mark.integration_saas
 @pytest.mark.integration_saas_override
-def test_mailchimp_override_access_request_task(
+async def test_mailchimp_override_access_request_task(
     db,
     policy,
     mailchimp_override_connection_config,
@@ -48,7 +48,7 @@ def test_mailchimp_override_access_request_task(
     merged_graph = mailchimp_override_dataset_config.get_graph()
     graph = DatasetGraph(merged_graph)
 
-    v = graph_task.run_access_request(
+    v = await graph_task.run_access_request(
         privacy_request,
         policy,
         graph,
@@ -140,7 +140,7 @@ def test_mailchimp_override_access_request_task(
 
 @pytest.mark.integration_saas
 @pytest.mark.integration_saas_override
-def test_mailchimp_erasure_request_task(
+async def test_mailchimp_erasure_request_task(
     db,
     policy,
     erasure_policy_string_rewrite,
@@ -161,7 +161,7 @@ def test_mailchimp_erasure_request_task(
     merged_graph = mailchimp_override_dataset_config.get_graph()
     graph = DatasetGraph(merged_graph)
 
-    graph_task.run_access_request(
+    await graph_task.run_access_request(
         privacy_request,
         policy,
         graph,
@@ -170,7 +170,7 @@ def test_mailchimp_erasure_request_task(
         db,
     )
 
-    v = graph_task.run_erasure(
+    v = await graph_task.run_erasure(
         privacy_request,
         erasure_policy_string_rewrite,
         graph,