+++ tests/ops/integration_tests/saas/test_datadog_task.py
@@ -21,7 +21,7 @@ def test_datadog_connection_test(datadog_connection_config) -> None:
 
 @pytest.mark.integration_saas
 @pytest.mark.integration_datadog
-def test_saas_access_request_task(
+async def test_saas_access_request_task(
     db,
     policy,
     datadog_connection_config,
@@ -42,7 +42,7 @@ def test_saas_access_request_task(
     dataset_name = datadog_connection_config.get_saas_config().fides_key
     merged_graph = datadog_dataset_config.get_graph()
     graph = DatasetGraph(merged_graph)
-    v = graph_task.run_access_request(
+    v = await graph_task.run_access_request(
         privacy_request,
         policy,
         graph,