+++ tests/ops/integration_tests/saas/test_hubspot_task.py
@@ -23,7 +23,7 @@ def test_hubspot_connection_test(connection_config_hubspot) -> None:
 
 @pytest.mark.integration_saas
 @pytest.mark.integration_hubspot
-def test_saas_access_request_task(
+async def test_saas_access_request_task(
     db,
     policy,
     connection_config_hubspot,
@@ -45,7 +45,7 @@ def test_saas_access_request_task(
     merged_graph = dataset_config_hubspot.get_graph()
     graph = DatasetGraph(merged_graph)
 
-    v = graph_task.run_access_request(
+    v = await graph_task.run_access_request(
         privacy_request,
         policy,
         graph,
@@ -128,7 +128,7 @@ def test_saas_access_request_task(
 
 @pytest.mark.integration_saas
 @pytest.mark.integration_hubspot
-def test_saas_erasure_request_task(
+async def test_saas_erasure_request_task(
     db,
     policy,
     erasure_policy_string_rewrite,
@@ -152,7 +152,7 @@ def test_saas_erasure_request_task(
     merged_graph = dataset_config_hubspot.get_graph()
     graph = DatasetGraph(merged_graph)
 
-    v = graph_task.run_access_request(
+    v = await graph_task.run_access_request(
         privacy_request, policy, graph, [connection_config_hubspot], identity_kwargs, db
     )
 
@@ -169,7 +169,7 @@ def test_saas_erasure_request_task(
 
     temp_masking = config.execution.masking_strict
     config.execution.masking_strict = False  # Allow delete
-    erasure = graph_task.run_erasure(
+    erasure = await graph_task.run_erasure(
         privacy_request,
         erasure_policy_string_rewrite,
         graph,